refactor(source-localization): split runner into modules
